+++ src/BaseAsset.php
@@ -17,55 +17,178 @@ abstract class BaseAsset implements Asset
         'url' => '',
         'handle' => '',
         'dependencies' => [],
+        'location' => Asset::FRONTEND,
         'version' => '',
         'enqueue' => true,
         'filters' => [],
-        'data' => [],
     ];
 
+    /**
+     * {@inheritDoc}
+     */
     public function url(): string
     {
-        return (string) $this->config['url'];
+        return (string) $this->config('url', '');
     }
 
+    /**
+     * {@inheritDoc}
+     */
     public function handle(): string
     {
-        return (string) $this->config['handle'];
+        return (string) $this->config('handle', '');
     }
 
-    public function dependencies(): array
-    {
-        return (array) $this->config['dependencies'];
-    }
-
+    /**
+     * {@inheritDoc}
+     */
     public function version(): string
     {
-        return (string) $this->config['version'];
+        return (string) $this->config('version', '');
     }
 
+    /**
+     * {@inheritDoc}
+     */
+    public function dependencies(): array
+    {
+        return array_unique($this->config('dependencies', []));
+    }
+
+    /**
+     * @param string ...$dependencies
+     *
+     * @return Script|Style
+     */
+    public function withDependencies(string ...$dependencies): Asset
+    {
+        $this->config['dependencies'] = array_merge(
+            $this->dependencies(),
+            $dependencies
+        );
+
+        return $this;
+    }
+
+    /**
+     * {@inheritDoc}
+     */
     public function location(): int
     {
-        return (int) ($this->config['location'] ?? self::FRONTEND);
+        return (int) $this->config('location', self::FRONTEND);
     }
 
+    /**
+     * @param int $location
+     *
+     * @return Script|Style
+     */
+    public function forLocation(int $location): Asset
+    {
+        $this->config['location'] = $location;
+
+        return $this;
+    }
+
+    /**
+     * {@inheritDoc}
+     */
     public function filters(): array
     {
-        return $this->config['filters'];
+        return $this->config('filters', []);
     }
 
-    public function data(): array
+    /**
+     * @param callable|OutputFilter\AssetOutputFilter ...$filters
+     *
+     * @return Style|Script
+     *
+     * // phpcs:disable Inpsyde.CodeQuality.ArgumentTypeDeclaration.NoArgumentType
+     */
+    public function withFilters(...$filters): Asset
     {
-        $data = $this->config['data'];
-        is_callable($data) and $data = $data();
+        $this->config['filters'] = array_merge(
+            $this->filters(),
+            $filters
+        );
 
-        return (array) $data;
+        return $this;
     }
 
+    /**
+     * {@inheritDoc}
+     */
     public function enqueue(): bool
     {
-        $enqueue = $this->config['enqueue'];
+        $enqueue = $this->config('enqueue', true);
         is_callable($enqueue) and $enqueue = $enqueue();
 
         return (bool) $enqueue;
+    }
+
+    /**
+     * {@inheritDoc}
+     *
+     * // phpcs:disable Inpsyde.CodeQuality.ArgumentTypeDeclaration.NoArgumentType
+     */
+    public function canEnqueue($enqueue): Asset
+    {
+        $this->config['enqueue'] = $enqueue;
+
+        return $this;
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    public function useHandler(string $handlerClass): Asset
+    {
+        $this->config['handler'] = $handlerClass;
+
+        return $this;
+    }
+
+    public function handler(): string
+    {
+        return (string) $this->config('handler', $this->defaultHandler());
+    }
+
+    /**
+     * @return string className of the default handler
+     */
+    abstract protected function defaultHandler(): string;
+
+    /**
+     * {@inheritDoc}
+     */
+    public function data(): array
+    {
+        return $this->config('data', []);
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    public function withCondition(string $condition): Asset
+    {
+        $this->config['data']['conditional'] = $condition;
+
+        return $this;
+    }
+
+    /**
+     * Retrieve a value from a config with a fallback if not existing.
+     *
+     * @param string $key
+     * @param null $default
+     *
+     * @return mixed|null
+     *
+     * // phpcs:disable Inpsyde.CodeQuality.ArgumentTypeDeclaration.NoArgumentType
+     * // phpcs:disable Inpsyde.CodeQuality.ReturnTypeDeclaration.NoReturnType
+     */
+    public function config(string $key, $default = null)
+    {
+        return $this->config[$key] ?? $default;
     }
 }