+++ src/Handler/ScriptHandler.php
@@ -13,6 +13,7 @@ namespace Inpsyde\Assets\Handler;
 use Inpsyde\Assets\Asset;
 use Inpsyde\Assets\OutputFilter\AsyncScriptOutputFilter;
 use Inpsyde\Assets\OutputFilter\DeferScriptOutputFilter;
+use Inpsyde\Assets\Script;
 
 class ScriptHandler implements AssetHandler, OutputFilterAwareAssetHandler
 {
@@ -48,6 +49,8 @@ class ScriptHandler implements AssetHandler, OutputFilterAwareAssetHandler
 
     public function register(Asset $asset): bool
     {
+        /** @var Script $asset */
+
         $handle = $asset->handle();
 
         wp_register_script(
@@ -57,10 +60,20 @@ class ScriptHandler implements AssetHandler, OutputFilterAwareAssetHandler
             $asset->version(),
             $asset->inFooter()
         );
+
         if (count($asset->localize()) > 0) {
             foreach ($asset->localize() as $name => $args) {
                 wp_localize_script($handle, $name, $args);
             }
+        }
+
+        foreach ($asset->inlineScripts() as $location => $data) {
+            wp_add_inline_script($handle, implode("\n", $data), $location);
+        }
+
+        $translation = $asset->translation();
+        if (isset($translation['domain']) && isset($translation['path'])) {
+            wp_set_script_translations($handle, $translation['domain'], $translation['path']);
         }
 
         if (count($asset->data()) > 0) {